// src/sets/types/IValidSetForSupplier.ts
import { IPosition } from '../../position/types/IPosition';

export interface IValidSetForSupplier {
  valid: boolean;
  setId: number;
  supplier: {
    id: number;
    company: string;
  };
  client: {
    id: number;
    company?: string;
    firstName: string;
    lastName: string;
  };
  positions: IPosition[];
}